adding base workflow class
