+++ sdk/java/spi/pom.xml
@@ -22,7 +22,7 @@
         <dependency>
             <groupId>io.cncf</groupId>
             <artifactId>serverlessworkflow-api</artifactId>
-            <version>${version}</version>
+            <version>${project.version}</version>
         </dependency>
 
         <!-- test -->
